# lib/api-zod/src/generated/types/equipoInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for TechStock inventory manager
 * OpenAPI spec version: 0.1.0
 */
import type { Categoria } from './categoria';
import type { EstadoEquipoCondicion } from './estadoEquipoCondicion';
import type { FormaPago } from './formaPago';

export interface EquipoInput {
  categoria: Categoria;
  /** @minLength 1 */
  equipo: string;
  /** @minLength 1 */
  marca: string;
  /** @minLength 1 */
  modelo: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  bateriaPct?: number;
  imeiSerial?: string;
  estadoEquipoCondicion: EstadoEquipoCondicion;
  fechaCompra: Date;
  proveedorId?: number;
  formaPagoCompra: FormaPago;
  /** @minimum 0 */
  precioCompra: number;
  /** @minimum 0 */
  gastosExtra?: number;
  comentarios?: string;
}