# lib/api-zod/src/generated/types/equipoUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for TechStock inventory manager
 * OpenAPI spec version: 0.1.0
 */
import type { Categoria } from './categoria';
import type { EstadoEquipoCondicion } from './estadoEquipoCondicion';
import type { FormaPago } from './formaPago';

export interface EquipoUpdate {
  categoria?: Categoria;
  /** @minLength 1 */
  equipo?: string;
  /** @minLength 1 */
  marca?: string;
  /** @minLength 1 */
  modelo?: string;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  bateriaPct?: number | null;
  /** @nullable */
  imeiSerial?: string | null;
  estadoEquipoCondicion?: EstadoEquipoCondicion;
  fechaCompra?: Date;
  /** @nullable */
  proveedorId?: number | null;
  formaPagoCompra?: FormaPago;
  /** @minimum 0 */
  precioCompra?: number;
  /** @minimum 0 */
  gastosExtra?: number;
  /** @nullable */
  comentarios?: string | null;
}